Skip duplicate downloads

+++ interproscan/modules/download/main.nf
@@ -8,7 +8,7 @@ process DOWNLOAD {
     label    'local', 'ips6_container'
 
     input:
-    tuple val(name), val(arcname), val(version)
+    tuple val(name), val(arcname), val(version), val(skip)
     val iprscan_version
     val outdir
 
@@ -16,15 +16,21 @@ process DOWNLOAD {
     tuple val(name), val(version), val("${outdir}/${arcname}/${version}")
 
     script:
-    """
-    cd ${outdir}
-    curl -OJ ${InterProScan.FTP_URL}/${iprscan_version}/${arcname}/${arcname}-${version}.tar.gz
-    curl -OJ ${InterProScan.FTP_URL}/${iprscan_version}/${arcname}/${arcname}-${version}.tar.gz.md5
-    md5sum -c ${arcname}-${version}.tar.gz.md5 || { echo "Error: MD5 checksum failed" >&2; exit 1; }
-    tar -zxf ${arcname}-${version}.tar.gz
-    rm ${arcname}-${version}.tar.gz*
-    chmod 777 -R ${arcname}
-    """
+    if (skip) {
+        """
+        """
+    } else {
+        """
+        cd ${outdir}
+        curl -OJ ${InterProScan.FTP_URL}/${iprscan_version}/${arcname}/${arcname}-${version}.tar.gz
+        curl -OJ ${InterProScan.FTP_URL}/${iprscan_version}/${arcname}/${arcname}-${version}.tar.gz.md5
+        md5sum -c ${arcname}-${version}.tar.gz.md5 || { echo "Error: MD5 checksum failed" >&2; exit 1; }
+        tar -zxf ${arcname}-${version}.tar.gz
+        rm ${arcname}-${version}.tar.gz*
+        chmod 777 -R ${arcname}
+        """        
+    }
+    
 }
 
 process FIND_MISSING_APP_DATA {
@@ -58,6 +64,7 @@ process FIND_MISSING_APP_DATA {
 
     with_data = [] as Set
     without_data = [] as Set
+    to_download = [] as Set
     apps_to_run.each { db_name ->
         if (app_dirs.containsKey(db_name)) {
             def normalised_name = db_name.replaceAll(/[\s\-]+/, '').toLowerCase()
@@ -70,7 +77,8 @@ process FIND_MISSING_APP_DATA {
             if (Files.exists(path)) {
                 with_data.add( [ normalised_name, db_version, path.toString() ])
             } else {
-                without_data.add( [ normalised_name, db_dir, db_version ] )
+                without_data.add( [ normalised_name, db_dir, db_version, to_download.contains(db_dir) ] )
+                to_download.add( db_dir )
             }
         }
     }
